Update typescript typing for validate function
